pyproject.toml
[project]
name = "Akashi"
requires-python = ">= 3.9"
version = "2.0.0"
dependencies = [
  "sqlalchemy[asyncio]==1.4.31",
  "pillow==9.0.1",
  "python-dateutil==2.8.1",
  "jishaku==2.5.1",
  "discord.py==2.1.*",
  "discord-ext-menus@git+https://github.com/Rapptz/discord-ext-menus",
  "discord-ext-menus-views@git+https://github.com/oliver-ni/discord-ext-menus-views",
  "discord-ext-ipc==2.0.0",
  "asyncpg==0.25.0",
  "prettytable==3.1.1",
  "pygit2==1.8.0",
  "psutil==5.9.0",
  "boto3==1.17.102",
  "aiofiles==0.7.0",
  "docutils==0.17.1",
  "toml==0.10.2",
  "humanize==3.13",
  "loguru==0.6.0",
  "alembic==1.7.6",
  "hondana",
  "lark",    
]
